Move Nil class from ast.core to ast.literals (#659)
