+++ pyccel/ast/literals.py
@@ -217,3 +217,11 @@ def get_default_literal_value(dtype):
         raise TypeError('Unknown type')
     return value
 
+class Nil(Basic):
+
+    """
+    class for None object in the code.
+    """
+
+    def __str__(self):
+        return 'None'
